NeuroComments/requirements.txt
Kurigram==2.1.36
g4f==0.4.1.0
python-dotenv==1.0.1
